Merging tidyscanner branch to make 1.5.0

+++ deprecated/TreeUtilityFunctions.R
@@ -326,7 +326,7 @@ patient.from.label <- function(label, regexp){
 
 read.count.from.label <- function(label, regexp){
   if(length(grep(regexp, label)>0)) {
-    return(as.numeric(sub(regexp, "\\3", label)))
+    return(as.integer(sub(regexp, "\\3", label)))
   } else {
     return(NA)
   }
